requirements.txt
pandas
confluent-kafka
python-dotenv
psycopg2-binary